Recipes sorted, damage multiplier in the config,
I think normal swords don't deal damage XD

+++ src/main/java/uk/co/wehavecookies56/kk/common/core/handler/EventHandler.java
@@ -1099,7 +1099,7 @@ public class EventHandler {
 
 			if (player.getHeldItem(EnumHand.MAIN_HAND) != null) {
 				if (player.getHeldItem(EnumHand.MAIN_HAND).getItem() instanceof ItemKeyblade) {
-					event.setAmount(event.getAmount() + ((ItemKeyblade) player.getHeldItem(EnumHand.MAIN_HAND).getItem()).getStrength() + player.getCapability(ModCapabilities.PLAYER_STATS, null).getStrength());
+					event.setAmount((float) (event.getAmount() + ((ItemKeyblade) player.getHeldItem(EnumHand.MAIN_HAND).getItem()).getStrength() + player.getCapability(ModCapabilities.PLAYER_STATS, null).getStrength()));
 					System.out.println("Old + Keyblade:" + event.getAmount());
 				}
 			}
